Refactor package of sql extension operator table
